+++ Assets/_Scripts/Interfaces/IDamageable.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(float dmg);
-    void Damage(float dmg, Vector3 dir);
+    void Damage(int dmg);
+    void Damage(int dmg, Vector3 dir);
 }
